completely remove hibernate.dialect.oracle.prefer_long_raw
Looks like this has not worked since probably H 6.0

The issue here is that the test for this feature was one of these tests
written to look at the Hibernate metamodel and not at the actual DDL that
was ultimately generated. This is why I despise "unit" testing instead of
testing actual user-visible functionality.

Anyway this was an undocumented feature so it can just be removed.