HHH-19080 Allow Hibernate Reactive to initialize the non lazy collections via PersistenceContext
Hibernate Reactive needs a reactive version of the PersistenceContext to
load non-lazy collections. But, now that `SatefulPersistenceContext` is
package-private, Hibernate Reactive cannot extend the class anymore.

These change make it possible to access the internal snapshots, so that the
reactive implementation can implement the interface and doesn't need to
extends `SatefulPersistenceContext`.