invert the default for hibernate.query.immutable_entity_update_query_handling_mode
The "default to warn but error when explicitly disabled" semantics for this were
entirely backwards and hostile to the user. "Safety" features should always be
on by default. Also, you all know what I think about WARNings, and here we had
the most annoying kind of WARN, the one that couldn't be suppressed, except in
log settings, even though it's probably intentional that the user is updating an
immutable entity. Ugh.