HHH-19497 Addresses major issue fallback implementation of "in list" predicate
When a NOT is applied to an IN, such as
cq.select(root).where(root.get("customerType").in(List.of(i1, i2)))
and when i1, i2 are represented by more than one column or formula (let's call these a and b), the previous implementation emitted:
(a <> ? OR b <> ?) OR (a <> ? OR b <> ?) OR ...
when the correct expression needs to be:
NOT ((a = ? AND b = ?) OR (a = ? AND b = ?))
So previous implementation incorrectly distributes the negation to each item in the list. De Morgan 101. The predicate is logically incorrect and this can lead to unpredictable results, which can even include ALL the records being returned from a table; or ZERO records being returned.