run the @OnDelete(CASCADE) tests also on Sybase
change the assertions about how many statements are executed

in the cases covered in these tests, the external behavior is
portable, which is great. On the other hand, there are also
cases where @OnDelete(CASCADE) is unportable, and perhaps in
those cases we should detect and throw when @OnDelete(CASCADE)
is used on a platform with no 'on delete cascade' construct
(i.e. on Sybase)