HHH-19565 @SQLRestriction should not be applied to a @ManyToOne
At least not when mapped to a foreign key association.

If we apply such a restriction, we will set the FK to null
when the entity is updated, resulting in data loss.

Note that this was the historical and correct behavior right
up until Hibernate 6.2. Some recent tests are asserting a
different and dangerous behavior, so I'm changing those tests.