HHH-19668 Reduce size of collections held in ParameterInterpretationImpl before storing them in the cache
Instances of ParameterInterpretationImpl are being held in the interpretation cache for native queries. We had reports of these elements having a substantial impact on memory consumption.

A simple improvement is to ensure the collections being held on by the container object are minimized; some more substantial improvements could also be investigated but I’ll leave those for another time.