HHH-20006 uniqueElement() did not have a self-consistent definition of "same"
Really, getSingleResult() is saying there should be a single element in
getResultList(). If that's not the case, you should use 'select distinct'.

That said, we might want a way to recover the old behavior (H5, apparently)
of making things distinct in memory according to equals(). Though, actually,
I believe you can just use a ResultListTransformer for that. Perhaps we
should just add a built-in one for migration purposes.